hugchat-bot/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
APScheduler==3.10.4
asgiref==3.8.1
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
frozenlist==1.4.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
persiantools==4.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
requests==2.31.0
six==1.16.0
tzlocal==5.2
urllib3==2.2.1
Werkzeug==3.0.2
yarl==1.9.4